requirements.txt
amqp
annotated-types
asgiref
billiard
celery
cffi
click
click-didyoumean
click-plugins
click-repl
contextlib2
cryptography
Django
django-cors-headers
django-jazzmin
django-ninja
django-ninja-extra
django-ninja-jwt
django-ninja-passwordreset
django-simple-history
dnspython
email_validator
h11
idna
injector
kombu
ninja-schema
prompt_toolkit
psycopg2
pycparser
pydantic
pydantic_core
PyJWT
python-dateutil
python-decouple
six
sqlparse
typing_extensions
tzdata
uvicorn
vine
wcwidth
